requirements.txt
playwright
openai
python-dotenv
pandas